change  src/app/cart/page.jsx  src/components/Cart.jsx: create a component and imported in the Cart/page.jsx
